+++ example/lib/ui/screen/view_interface.dart
@@ -1,0 +1,7 @@
+import 'package:mvvm_flutter/base_interface/base_layer.dart';
+
+///used in <a href="photo_viewmodel.dart">PhotoViewModel.class</a>
+abstract class IPhotoView extends IBaseView {
+  void onChangeBackground();
+  void onLoadImageData();
+}