+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-## 0.0.1
+## [0.1.0] - 
+[![pub package](https://img.shields.io/pub/v/mvvm_flutter.svg)](https://pub.dev/packages/mvvm_flutter)
+[![join chat](https://badges.gitter.im/unicreators/mvvm_flutter.svg)](https://gitter.im/unicreators/mvvm_flutter)
 
-* TODO: Describe initial release.
+
+
+A Flutter MVVM (Model-View-ViewModel) implementation. 
+It uses abstraction layer between flutter view and model class .
+ binding to establish a connection between the ViewModel and the View 
+ , and drives the View changes through the ViewModel.
+  
+ 
+